data_persistence_app/requirements.txt
kafka-python
psycopg2
requests
datetime